# backend/docs/swagger.yaml
basePath: /
definitions:
  auth.CurrentUserCompany:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  auth.CurrentUserCompanyMember:
    properties:
      display_name:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
  auth.CurrentUserResponse:
    properties:
      company:
        $ref: '#/definitions/auth.CurrentUserCompany'
      user:
        $ref: '#/definitions/auth.CurrentUserCompanyMember'
    type: object
  auth.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  auth.LoginCompany:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      access_token:
        type: string
      company:
        $ref: '#/definitions/auth.LoginCompany'
      expires_in:
        type: integer
      token_type:
        type: string
      user:
        $ref: '#/definitions/auth.LoginUser'
    type: object
  auth.LoginUser:
    properties:
      display_name:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
  company.CreateRequest:
    properties:
      name:
        maxLength: 150
        minLength: 2
        type: string
      slug:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    - slug
    type: object
  company.CreateResponse:
    properties:
      company_id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
  company.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  user.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  user.RegisterRequest:
    properties:
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      password_confirmation:
        type: string
    required:
    - display_name
    - email
    - password
    - password_confirmation
    type: object
  user.RegisterResponse:
    properties:
      company_id:
        type: integer
      display_name:
        type: string
      email:
        type: string
      role:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8081
info:
  contact: {}
  description: FlowSync Pulse バックエンドAPI仕様書
  title: FlowSync Pulse API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: メールアドレスとパスワードで認証し、アクセストークンとユーザー情報を返します。
      parameters:
      - description: ログイン情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: ログイン
      tags:
      - auth
  /api/companies:
    post:
      consumes:
      - application/json
      description: 会社名と会社スラッグを指定して、新しい会社を登録します。
      parameters:
      - description: 会社情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/company.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/company.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/company.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/company.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/company.ErrorResponse'
      summary: 会社登録
      tags:
      - companies
  /api/companies/{companyId}/users:
    post:
      consumes:
      - application/json
      description: 指定した会社に所属するユーザーを登録します。
      parameters:
      - description: 会社ID
        in: path
        name: companyId
        required: true
        type: integer
      - description: ユーザー情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.ErrorResponse'
      summary: ユーザー登録
      tags:
      - users
  /api/health:
    get:
      description: APIサーバーとデータベースの接続状態を返します。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ヘルスチェック
      tags:
      - health
  /api/me:
    get:
      description: アクセストークンから現在のログインユーザーと会社情報を取得します。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.CurrentUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ログインユーザー取得
      tags:
      - auth
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"